framework_requirements.txt
# CAD Data Correction Framework - Python Dependencies
# Install with: pip install -r framework_requirements.txt

# Core data processing
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.10          # Excel file support

# Configuration
PyYAML>=6.0

# System monitoring
psutil>=5.9.0

# Data validation
pydantic>=2.0.0

# Optional: Large dataset processing
dask[complete]>=2023.1.0

# Optional: Progress bars
tqdm>=4.65.0

# Development and testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
